renamed files to be in snake_case
